backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# LangChain and LangGraph
langgraph==0.2.52
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7

# Web scraping and data ingestion
beautifulsoup4==4.12.2
selenium==4.16.0
scrapy==2.11.0
requests==2.31.0
linkedin-api==2.2.0

# Vector database for memory
chromadb==0.4.22
sentence-transformers==2.2.2

# Additional utilities
celery==5.3.4
redis==5.0.1
python-linkedin-v2==0.9.1

